Adding some demos. Getting pissed off.
My lower common tangent system works, but doesn't work well with points with the same x-coord. I'm scrambling their positions a bit to try to hack my way out of it.

There's all sorts of tangling, though, that just ruins the mesh. I'll have to figure out how to deal with that.